Add `isReadonly` flag to Array and Record runtypes
The flag causes the static type to change. For Record the static type
has all properties marked readonly, and for Array the result type is the
TS ReadonlyArray (or `readonly T[]` in the upcoming 3.4 version).

Both runtypes also have a convenience method, `asReadonly`, which
returns the same type but with the `isReadonly` flag set to true. Some
users may prefer the method call over setting the flag for esthetic
reasons.